# Conflicts:
#	templates/workspaces/base/porter.yaml
#	templates/workspaces/base/template_schema.json

+++ api_app/db/migrations/shared_services.py
@@ -2,6 +2,8 @@ import logging
 
 from azure.cosmos import CosmosClient
 from db.repositories.shared_services import SharedServiceRepository
+from db.repositories.resources import IS_OPERATING_SHARED_SERVICE
+import semantic_version
 
 
 class SharedServiceMigration(SharedServiceRepository):
@@ -13,10 +15,13 @@ class SharedServiceMigration(SharedServiceRepository):
 
         migrated = False
         for template_name in template_names:
-            for item in self.query(query=f'SELECT * FROM c WHERE c.resourceType = "shared-service" AND c.templateName = "{template_name}" \
-                                                                ORDER BY c.updatedWhen ASC OFFSET 1 LIMIT 10000'):
-                logging.info(f'Deleting element {item["id"]}')
-                self.delete_item(item["id"])
-                migrated = True
+            for item in self.query(query=f'SELECT * FROM c WHERE c.resourceType = "shared-service" \
+                                           AND c.templateName = "{template_name}" AND {IS_OPERATING_SHARED_SERVICE} \
+                                           ORDER BY c.updatedWhen ASC OFFSET 1 LIMIT 10000'):
+                template_version = semantic_version.Version(item["templateVersion"])
+                if (template_version < semantic_version.Version('0.3.0')):
+                    logging.info(f'Deleting element {item["id"]}')
+                    self.delete_item(item["id"])
+                    migrated = True
 
         return migrated
